move GET tag endpoints to minimal apis
